+++ common/src/main/java/nl/enjarai/a_good_place/mixin/RebuildTaskMixin.java
@@ -2,14 +2,14 @@ package nl.enjarai.a_good_place.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.sugar.Local;
-import net.minecraft.client.renderer.chunk.ChunkRenderDispatcher;
+import net.minecraft.client.renderer.chunk.SectionCompiler;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.RenderShape;
 import nl.enjarai.a_good_place.particles.BlocksParticlesManager;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
-@Mixin(ChunkRenderDispatcher.RenderChunk.RebuildTask.class)
+@Mixin(SectionCompiler.class)
 public abstract class RebuildTaskMixin {
 
     @ModifyExpressionValue(
@@ -22,6 +22,20 @@ public abstract class RebuildTaskMixin {
     private RenderShape wonkyblock$hideBlock(RenderShape original, @Local(ordinal = 2) BlockPos blockPos) {
         if (original != RenderShape.INVISIBLE && BlocksParticlesManager.isBlockHidden(blockPos)) {
             return RenderShape.INVISIBLE;
+        }
+        return original;
+    }
+
+    @ModifyExpressionValue(
+            method = "compile",
+            at = @At(
+                    value = "INVOKE",
+                    target = "Lnet/minecraft/world/level/block/state/BlockState;isSolidRender()Z"
+            )
+    )
+    private boolean wonkyblock$overrideSolidRender(boolean original, @Local(ordinal = 2) BlockPos blockPos) {
+        if (original && BlocksParticlesManager.isBlockHidden(blockPos)) {
+            return false;
         }
         return original;
     }