+++ common/src/main/java/nl/enjarai/a_good_place/pack/state_tests/AnyOf.java
@@ -1,6 +1,7 @@
 package nl.enjarai.a_good_place.pack.state_tests;
 
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.MapCodec;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
@@ -9,8 +10,8 @@ import java.util.List;
 
 public record AnyOf(List<BlockStatePredicate> predicates) implements BlockStatePredicate {
 
-    public static final Codec<AnyOf> CODEC = BlockStatePredicate.CODEC.listOf()
-            .fieldOf("predicates").xmap(AnyOf::new, AnyOf::predicates).codec();
+    public static final MapCodec<AnyOf> CODEC = BlockStatePredicate.CODEC.listOf()
+            .fieldOf("predicates").xmap(AnyOf::new, AnyOf::predicates);
 
     @Override
     public BlockStatePredicateType<?> getType() {