+++ common/src/main/java/nl/enjarai/a_good_place/pack/state_tests/Not.java
@@ -1,14 +1,15 @@
 package nl.enjarai.a_good_place.pack.state_tests;
 
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.MapCodec;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 
 public record Not(BlockStatePredicate predicate) implements BlockStatePredicate {
 
-    public static final Codec<Not> CODEC = BlockStatePredicate.CODEC
-            .fieldOf("predicate").xmap(Not::new, Not::predicate).codec();
+    public static final MapCodec<Not> CODEC = BlockStatePredicate.CODEC
+            .fieldOf("predicate").xmap(Not::new, Not::predicate);
 
     @Override
     public BlockStatePredicateType<?> getType() {