+++ common/src/main/java/nl/enjarai/a_good_place/particles/BlocksParticlesManager.java
@@ -1,11 +1,9 @@
 package nl.enjarai.a_good_place.particles;
 
-import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.Camera;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
-import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.InteractionHand;
@@ -43,7 +41,7 @@ public class BlocksParticlesManager {
         if (param != null) {
             Camera camera = Minecraft.getInstance().gameRenderer.getMainCamera();
 
-            if (camera.getPosition().distanceToSqr(pos.getCenter()) <= 1024.0) {
+            if (camera.position().distanceToSqr(pos.getCenter()) <= 1024.0) {
                 var p = new ConfiguredPlacingParticle(level, pos, face, player, hand, param);
 
                 var old = PARTICLES.put(pos, p);
@@ -107,34 +105,13 @@ public class BlocksParticlesManager {
     }
 
     public static void renderParticles(PoseStack poseStack, float tickDelta) {
-        if (AGoodPlace.RENDER_AS_VANILLA_PARTICLES || PARTICLES.isEmpty()) return;
-
-        poseStack.pushPose();
-
-        Minecraft mc = Minecraft.getInstance();
-        Camera camera = mc.gameRenderer.getMainCamera();
-        MultiBufferSource.BufferSource bufferSource = mc.renderBuffers().bufferSource();
-
-        // lightTexture.turnOnLightLayer();
-        RenderSystem.enableDepthTest();
-        PoseStack poseStack2 = RenderSystem.getModelViewStack();
-        poseStack2.pushPose();
-        poseStack2.mulPoseMatrix(poseStack.last().pose());
-        RenderSystem.applyModelViewMatrix();
-
+        if (PARTICLES.isEmpty()) return;
+        Camera camera = Minecraft.getInstance().gameRenderer.getMainCamera();
         for (var p : PARTICLES.values()) {
-            p.render(null, camera, tickDelta);
+            if (p.isAlive()) {
+                p.render(camera, tickDelta);
+            }
         }
-        bufferSource.endBatch();
-
-
-        poseStack2.popPose();
-        RenderSystem.applyModelViewMatrix();
-        RenderSystem.depthMask(true);
-        RenderSystem.disableBlend();
-        // lightTexture.turnOffLightLayer();
-
-        poseStack.popPose();
     }
 
     public static void modifyTilePosition(BlockPos pos, PoseStack pose, float partialTicks) {