+++ forge/src/main/java/nl/enjarai/a_good_place/forge/AGoodPlaceImpl.java
@@ -1,51 +1,48 @@
 package nl.enjarai.a_good_place.forge;
 
 import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.blaze3d.vertex.VertexConsumer;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.block.BlockRenderDispatcher;
 import net.minecraft.client.renderer.texture.OverlayTexture;
-import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
+import net.minecraft.server.packs.PackLocationInfo;
+import net.minecraft.server.packs.PackResources;
+import net.minecraft.server.packs.PackSelectionConfig;
 import net.minecraft.server.packs.PackType;
-import net.minecraft.server.packs.metadata.pack.PackMetadataSection;
+import net.minecraft.server.packs.PathPackResources;
 import net.minecraft.server.packs.repository.Pack;
 import net.minecraft.server.packs.repository.PackSource;
 import net.minecraft.server.packs.resources.PreparableReloadListener;
-import net.minecraft.util.RandomSource;
-import net.minecraft.world.flag.FeatureFlagSet;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.RenderTypeHelper;
-import net.minecraftforge.client.event.ClientPlayerNetworkEvent;
-import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
-import net.minecraftforge.client.event.RenderLevelStageEvent;
-import net.minecraftforge.client.model.data.ModelData;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.data.loading.DatagenModLoader;
-import net.minecraftforge.event.AddPackFindersEvent;
-import net.minecraftforge.event.TickEvent;
-import net.minecraftforge.event.level.LevelEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.ModList;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.fml.loading.FMLEnvironment;
-import net.minecraftforge.fml.loading.FMLLoader;
-import net.minecraftforge.forgespi.locating.IModFile;
-import net.minecraftforge.resource.PathPackResources;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.ModList;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
+import net.neoforged.fml.loading.FMLEnvironment;
+import net.neoforged.neoforge.client.event.AddClientReloadListenersEvent;
+import net.neoforged.neoforge.client.event.ClientPlayerNetworkEvent;
+import net.neoforged.neoforge.client.event.RenderLevelStageEvent;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.data.loading.DatagenModLoader;
+import net.neoforged.neoforge.event.AddPackFindersEvent;
+import net.neoforged.neoforge.event.level.LevelEvent;
+import net.neoforged.neoforge.event.tick.LevelTickEvent;
+import net.neoforged.neoforgespi.locating.IModFile;
 import nl.enjarai.a_good_place.AGoodPlace;
 import nl.enjarai.a_good_place.pack.AnimationsManager;
 import nl.enjarai.a_good_place.pack.state_tests.BlockStatePredicateType;
 import nl.enjarai.a_good_place.particles.BlocksParticlesManager;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.Objects;
+import java.nio.file.Path;
+import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
@@ -54,22 +51,21 @@ import java.util.function.Supplier;
 public class AGoodPlaceImpl {
     public static final String MOD_ID = AGoodPlace.MOD_ID;
 
-    public AGoodPlaceImpl() {
-        if (FMLEnvironment.dist == Dist.CLIENT) {
+    public AGoodPlaceImpl(IEventBus modEventBus) {
+        if (FMLEnvironment.getDist() == Dist.CLIENT) {
 
-            FMLJavaModLoadingContext.get().getModEventBus()
-                            .addListener(this::onSetup);
+            modEventBus.addListener(this::onSetup);
 
-            addClientReloadListener(AnimationsManager::new, AGoodPlace.res("animations"));
+            addClientReloadListener(AnimationsManager::new, AGoodPlace.res("animations"), modEventBus);
 
             boolean firstInstall = AGoodPlace.copySamplePackIfNotPresent();
-            MinecraftForge.EVENT_BUS.register(this);
+            NeoForge.EVENT_BUS.register(this);
 
             registerOptionalTexturePack(AGoodPlace.res("default_animations"),
-                    Component.nullToEmpty("Default Place Animations"), firstInstall);
+                    Component.nullToEmpty("Default Place Animations"), firstInstall, modEventBus);
 
             BlockStatePredicateType.init();
-            AGoodPlace.IS_DEV = !FMLLoader.isProduction();
+            AGoodPlace.IS_DEV = !FMLEnvironment.isProduction();
         }
     }
 
@@ -90,66 +86,56 @@ public class AGoodPlaceImpl {
     }
 
     @SubscribeEvent
-    public void onRenderWorld(RenderLevelStageEvent event) {
-        if (event.getStage() == RenderLevelStageEvent.Stage.AFTER_ENTITIES) {
-            BlocksParticlesManager.renderParticles(event.getPoseStack(), event.getPartialTick());
-        }
+    public void onRenderWorld(RenderLevelStageEvent.AfterEntities event) {
+        BlocksParticlesManager.renderParticles(event.getPoseStack(), Minecraft.getInstance().getDeltaTracker().getGameTimeDeltaPartialTick(false));
     }
 
     @SubscribeEvent
-    public void onClientTick(TickEvent.LevelTickEvent tickEvent) {
-        if (tickEvent.phase == TickEvent.Phase.END && tickEvent.level.isClientSide) {
-            BlocksParticlesManager.tickParticles((ClientLevel) tickEvent.level);
+    public void onClientTick(LevelTickEvent.Post tickEvent) {
+        if (tickEvent.getLevel().isClientSide()) {
+            BlocksParticlesManager.tickParticles((ClientLevel) tickEvent.getLevel());
         }
     }
 
-    public static void renderBlock(BakedModel model, long seed, PoseStack poseStack, MultiBufferSource buffer, BlockState state,
+    public static void renderBlock(PoseStack poseStack, MultiBufferSource buffer, BlockState state,
                                    Level level, BlockPos pos, BlockRenderDispatcher dispatcher) {
-        //same as ForgeHooksClient.renderPistonMovedBlocks (what pistons use)
-        for (var renderType : model.getRenderTypes(state, RandomSource.create(seed), ModelData.EMPTY)) {
-            VertexConsumer vertexConsumer = buffer.getBuffer(RenderTypeHelper.getMovingBlockRenderType(renderType));
-            dispatcher.getModelRenderer().tesselateBlock(level, model, state, pos, poseStack, vertexConsumer, false,
-                    RandomSource.create(), state.getSeed(pos), OverlayTexture.NO_OVERLAY, ModelData.EMPTY, renderType);
-        }
+        dispatcher.renderSingleBlock(state, poseStack, buffer, 0xF000F0, OverlayTexture.NO_OVERLAY);
     }
 
-    public static void addClientReloadListener(Supplier<PreparableReloadListener> listener, ResourceLocation location) {
-        Consumer<RegisterClientReloadListenersEvent> eventConsumer = (event) -> {
-            event.registerReloadListener(listener.get());
+    public static void addClientReloadListener(Supplier<PreparableReloadListener> listener, Identifier location, IEventBus modEventBus) {
+        Consumer<AddClientReloadListenersEvent> eventConsumer = (event) -> {
+            event.addListener(location, listener.get());
         };
-        FMLJavaModLoadingContext.get().getModEventBus().addListener(eventConsumer);
+        modEventBus.addListener(eventConsumer);
     }
 
-    public static void registerOptionalTexturePack(ResourceLocation folderName, Component displayName, boolean defaultEnabled) {
+    public static void registerOptionalTexturePack(Identifier folderName, Component displayName, boolean defaultEnabled, IEventBus modEventBus) {
         registerResourcePack(PackType.CLIENT_RESOURCES,
                 () -> {
-                    IModFile file = ModList.get().getModFileById(folderName.getNamespace()).getFile();
-                    try (PathPackResources pack = new PathPackResources(
-                            folderName.toString(),
-                            true,
-                            file.findResource("resourcepacks/" + folderName.getPath()))) {
-                        var metadata = Objects.requireNonNull(pack.getMetadataSection(PackMetadataSection.TYPE));
-                        return Pack.create(
-                                folderName.toString(),
-                                displayName,
-                                defaultEnabled,
-                                (s) -> pack,
-                                new Pack.Info(metadata.getDescription(), metadata.getPackFormat(), FeatureFlagSet.of()),
-                                PackType.CLIENT_RESOURCES,
-                                Pack.Position.TOP,
-                                false,
-                                PackSource.BUILT_IN);
+                    try {
+                        IModFile file = ModList.get().getModFileById(folderName.getNamespace()).getFile();
+                        Path contentRoot = file.getContents().getContentRoots().iterator().next();
+                        Path packPath = contentRoot.resolve("resourcepacks").resolve(folderName.getPath());
+                        PackLocationInfo info = new PackLocationInfo(folderName.toString(), displayName, PackSource.BUILT_IN, Optional.empty());
+                        Pack.ResourcesSupplier supplier = new Pack.ResourcesSupplier() {
+                            @Override
+                            public PackResources openPrimary(PackLocationInfo i) { return new PathPackResources(i, packPath); }
+                            @Override
+                            public PackResources openFull(PackLocationInfo i, Pack.Metadata m) { return new PathPackResources(i, packPath); }
+                        };
+                        PackSelectionConfig selectionConfig = new PackSelectionConfig(defaultEnabled, Pack.Position.TOP, false);
+                        return Pack.readMetaAndCreate(info, supplier, PackType.CLIENT_RESOURCES, selectionConfig);
                     } catch (Exception ee) {
                         if (!DatagenModLoader.isRunningDataGen()) ee.printStackTrace();
                     }
                     return null;
-                }
+                },
+                modEventBus
         );
     }
 
-    public static void registerResourcePack(PackType packType, @Nullable Supplier<Pack> packSupplier) {
+    public static void registerResourcePack(PackType packType, @Nullable Supplier<Pack> packSupplier, IEventBus modEventBus) {
         if (packSupplier == null) return;
-        var bus = FMLJavaModLoadingContext.get().getModEventBus();
         Consumer<AddPackFindersEvent> consumer = event -> {
             if (event.getPackType() == packType) {
                 var p = packSupplier.get();
@@ -158,7 +144,7 @@ public class AGoodPlaceImpl {
                 }
             }
         };
-        bus.addListener(consumer);
+        modEventBus.addListener(consumer);
     }